+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+### Changed
+
+- upgraded babel, webpack and other relevant packages
+
 ## [0.8.1] - 2018-10-10
 
 ### Fixed